Build 0008
• Added method SettingsIO. deleteSettingsINI() to delete settings.ini
• Updated PSTInterface.getOutlookFolders() so that Outlook 2003 folders
can be found (there are still issues with daylight savings time. As a
result, some appointments in the google calendar are off by an hour)
• If the nimbus look and feel can't be located, the default look and
feel is now used instead of being set to null
• Updates to javadocs
• Made com.korshyadoo.calendar.PSTInterface.pstLocation public final
instead of private
• Removed com.korshyadoo.calendar.PSTInterface.getPSTLocation()
• Made com.korshyadoo.calendar.PSTInterface a singleton
• Made com.korshyadoo.calendar.SettingsIO a singleton
• Updated MainFrame to accomodate com.korshyadoo.calendar.PSTInterface
now being a singleton
• Added com.korshyadoo.calendar.PSTInterface.refresh()
• updates com.korshyadoo.calendar.PSTInterface.pstLocation to what is
stored in the memory buffer of com.korshyadoo.calendar.SettingsIO
• Redesigned com.korshyadoo.calendar.SettingsIO
• renamed settings field to settingsBuffer
• removed checkSettingsState() and replaced with createFile() and
settingsExists()
• Changed return on readSettings to StringBuilder so it can be used to
initialize settingsBuffer whether or not settings.ini is empty
• Renamed isEmpty() to isEmptyBuffer()
• Changing users will now set the success label, the success checkmark,
the action time, and the number of events to not visible
• The message displayed while searching for the .pst file is now hidden
when multiple results are shown
• While searching for the .pst file, a label will now show which
directory is currently being searched and the mouse pointer changes

• Bug Fix: When doing a sync, if Outlook.pst was not found in the
location specified by settings.ini the message indicated that the user
did not have permission to access the file
• Bug Fix: After a search is made for the .pst file and the user closes
the login page, the next launch will cause an exception.